FIX #4: Support IAM role-based access to S3 bucket (#5)
This patch changes s3fs.py and s3manager.py to allow 'None' as valid values
for 'access_key_id' and 'secret_access_key'. When using IAM role-based access to S3
buckets, the application must not specify any value for these two items at all.

The patch also changes the default value of these configuration items to 'None', as
that can actually be a useful value (the previous default, an empty string, would never
be an useful value). With this change, if the user fails to configure one (or both)
of these values, the error they receive will indicate that no value was supplied,
instead of indicating that the Access Key ID could not be found (or that the Secret
Access Key is invalid).

This fixes issue #4.